+++ changelog.md
@@ -11,6 +11,7 @@
 ## Improvements
 
 - Support `neo4j.EagerResult` in the `from_neo4j` integration which is the default return type by `neo4j.Driver.execute_query()`.
+- Detect light/dark theme changes and adapt rendering unless theme was explicitly set. Before the theme would only be checked on the first render.
 
 
 ## Other changes